Refactored input images
